rename gamelist to list_game
rename list_music_album to list_music
