+++ classes/list_game.rb
