+++ store/books.json
@@ -1,1 +1,1 @@
-[{"id":335,"publisher":"welshire Publishers","publish_date":"2022-02-02","cover_state":"Good"},{"id":221,"publisher":"New york times","publish_date":"2022-07-07","cover_state":"GOOD"},{"id":384,"publisher":"Chicago Publishers","publish_date":"2000-05-08","cover_state":"GOOD"},{"id":79,"publisher":"Mccmillan Publishers","publish_date":"2020-05-07","cover_state":"GOOD"},{"id":863,"publisher":"New York Times Publishers","publish_date":"2022-05-05","cover_state":"GOOD"},{"id":392,"publisher":"Dannish Publishers","publish_date":"2022-09-09","cover_state":"GOOD"},{"id":463,"publisher":"New york","publish_date":"2002-06-06","cover_state":"GOOD"},{"id":589,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":675,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":639,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":444,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":273,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":131,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":293,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":530,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":850,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":997,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":687,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":938,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"}]
+[{"id":335,"publisher":"welshire Publishers","publish_date":"2022-02-02","cover_state":"Good"},{"id":221,"publisher":"New york times","publish_date":"2022-07-07","cover_state":"GOOD"},{"id":384,"publisher":"Chicago Publishers","publish_date":"2000-05-08","cover_state":"GOOD"},{"id":79,"publisher":"Mccmillan Publishers","publish_date":"2020-05-07","cover_state":"GOOD"},{"id":863,"publisher":"New York Times Publishers","publish_date":"2022-05-05","cover_state":"GOOD"},{"id":392,"publisher":"Dannish Publishers","publish_date":"2022-09-09","cover_state":"GOOD"},{"id":463,"publisher":"New york","publish_date":"2002-06-06","cover_state":"GOOD"},{"id":589,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":675,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":639,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":444,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":273,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":131,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":293,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":530,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":850,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":997,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":687,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":938,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"},{"id":539,"publisher":"Publisher","publish_date":"2023-06-07","cover_state":"GOOD"}]
