+++ store/labels.json
@@ -1,1 +1,1 @@
-[{"id":722,"title":"The cure","color":"red"},{"id":688,"title":"The Shire","color":"White"},{"id":457,"title":"Hunger Games","color":"Brown"},{"id":52,"title":"Harry Potter and the Goblet of Fire","color":"Red"},{"id":371,"title":"The Recruit","color":"Blue"},{"id":713,"title":"The Frozen","color":"Pink"},{"id":567,"title":"The Joker","color":"Red"},{"id":487,"title":"Title","color":"Color"},{"id":455,"title":"Title","color":"Color"},{"id":439,"title":"Title","color":"Color"},{"id":658,"title":"Title","color":"Color"},{"id":933,"title":"Title","color":"Color"},{"id":540,"title":"Title","color":"Color"},{"id":250,"title":"Title","color":"Color"},{"id":299,"title":"Title","color":"Color"},{"id":883,"title":"Title","color":"Color"},{"id":158,"title":"Title","color":"Color"},{"id":246,"title":"Title","color":"Color"},{"id":730,"title":"Title","color":"Color"}]
+[{"id":722,"title":"The cure","color":"red"},{"id":688,"title":"The Shire","color":"White"},{"id":457,"title":"Hunger Games","color":"Brown"},{"id":52,"title":"Harry Potter and the Goblet of Fire","color":"Red"},{"id":371,"title":"The Recruit","color":"Blue"},{"id":713,"title":"The Frozen","color":"Pink"},{"id":567,"title":"The Joker","color":"Red"},{"id":487,"title":"Title","color":"Color"},{"id":455,"title":"Title","color":"Color"},{"id":439,"title":"Title","color":"Color"},{"id":658,"title":"Title","color":"Color"},{"id":933,"title":"Title","color":"Color"},{"id":540,"title":"Title","color":"Color"},{"id":250,"title":"Title","color":"Color"},{"id":299,"title":"Title","color":"Color"},{"id":883,"title":"Title","color":"Color"},{"id":158,"title":"Title","color":"Color"},{"id":246,"title":"Title","color":"Color"},{"id":730,"title":"Title","color":"Color"},{"id":63,"title":"Title","color":"Color"}]
